+++ android/src/main/kotlin/com/klippa/scanner/klippa_scanner_sdk/KlippaScannerSdkPlugin.kt
@@ -3,17 +3,21 @@ package com.klippa.scanner.klippa_scanner_sdk
 
 import android.app.Activity
 import android.content.Context
-import android.content.Intent
-import android.media.Image
 import android.util.Log
 import android.util.Size
-import androidx.annotation.NonNull
 import com.klippa.scanner.KlippaScannerBuilder
 import com.klippa.scanner.KlippaScannerListener
-import com.klippa.scanner.model.KlippaImage
+import com.klippa.scanner.model.Failure
+import com.klippa.scanner.model.Instructions
+import com.klippa.scanner.model.KlippaCameraModes
+import com.klippa.scanner.model.KlippaDocumentMode
 import com.klippa.scanner.model.KlippaImageColor
+import com.klippa.scanner.model.KlippaMultipleDocumentMode
 import com.klippa.scanner.model.KlippaObjectDetectionModel
 import com.klippa.scanner.model.KlippaScannerResult
+import com.klippa.scanner.model.KlippaSegmentedDocumentMode
+import com.klippa.scanner.model.KlippaSingleDocumentMode
+import com.klippa.scanner.model.Success
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
@@ -21,7 +25,6 @@ import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
-import io.flutter.plugin.common.PluginRegistry
 
 /** KlippaScannerSdkPlugin */
 class KlippaScannerSdkPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
@@ -35,15 +38,17 @@ class KlippaScannerSdkPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
 
   private var activityPluginBinding : ActivityPluginBinding? = null
 
-  private val SESSION_REQUEST_CODE = 9293
   private val E_ACTIVITY_DOES_NOT_EXIST = "E_ACTIVITY_DOES_NOT_EXIST"
   private val E_MISSING_SESSION_TOKEN = "E_MISSING_SESSION_TOKEN"
   private val E_FAILED_TO_SHOW_SESSION = "E_FAILED_TO_SHOW_SESSION"
   private val E_CANCELED = "E_CANCELED"
-  private val E_UNKNOWN_ERROR = "E_UNKNOWN_ERROR"
   private var resultHandler : Result? = null
 
-  val listener = object : KlippaScannerListener {
+  private var singleDocumentModeInstructionsDismissed = false
+  private var multiDocumentModeInstructionsDismissed = false
+  private var segmentedDocumentModeInstructionsDismissed = false
+
+  private val listener = object : KlippaScannerListener {
 
     override fun klippaScannerDidFinishScanningWithResult(result: KlippaScannerResult) {
       val images: MutableList<Map<String, String>> = mutableListOf()
@@ -53,15 +58,17 @@ class KlippaScannerSdkPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
         images.add(imageDict)
       }
 
-      val multipleDocuments: Boolean = result.multipleDocumentsModeEnabled
       val cropEnabled: Boolean = result.cropEnabled
       val timerEnabled: Boolean = result.timerEnabled
 
       val resultDict = mapOf(
         "Images" to images,
-        "MultipleDocuments" to multipleDocuments,
         "Crop" to cropEnabled,
-        "TimerEnabled" to timerEnabled)
+        "TimerEnabled" to timerEnabled,
+        "SingleDocumentModeInstructionsDismissed" to singleDocumentModeInstructionsDismissed,
+        "MultiDocumentModeInstructionsDismissed" to multiDocumentModeInstructionsDismissed,
+        "SegmentedDocumentModeInstructionsDismissed" to segmentedDocumentModeInstructionsDismissed
+        )
 
       resultHandler?.success(resultDict)
       resultHandler = null
@@ -77,7 +84,7 @@ class KlippaScannerSdkPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
 
   }
 
-  override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+  override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
     channel = MethodChannel(flutterPluginBinding.binaryMessenger, "klippa_scanner_sdk")
     channel.setMethodCallHandler(this)
     context = flutterPluginBinding.applicationContext
@@ -101,7 +108,7 @@ class KlippaScannerSdkPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
 
 
 
-  override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
+  override fun onMethodCall(call: MethodCall, result: Result) {
     if (call.method == "startSession") {
       startSession(call, result)
     } else {
@@ -109,7 +116,7 @@ class KlippaScannerSdkPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
     }
   }
 
-  private fun startSession(@NonNull call: MethodCall, @NonNull result: Result) {
+  private fun startSession(call: MethodCall, result: Result) {
     if (activityPluginBinding == null) {
       result.error(E_ACTIVITY_DOES_NOT_EXIST, "Activity doesn't exist", null)
       return
@@ -122,14 +129,6 @@ class KlippaScannerSdkPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
       }
 
       val builder = KlippaScannerBuilder(listener, call.argument<String>("License")!!)
-
-      if (call.hasArgument("AllowMultipleDocuments")) {
-        builder.menu.allowMultiDocumentsMode = call.argument<Boolean>("AllowMultipleDocuments")!!
-      }
-
-      if (call.hasArgument("DefaultMultipleDocuments")) {
-        builder.menu.isMultiDocumentsModeEnabled = call.argument<Boolean>("DefaultMultipleDocuments")!!
-      }
 
       if (call.hasArgument("DefaultCrop")) {
         builder.menu.isCropEnabled = call.argument<Boolean>("DefaultCrop")!!
@@ -198,7 +197,7 @@ class KlippaScannerSdkPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
         builder.messages.successMessage = call.argument<String>("Success.message")!!
       }
 
-      if (call.hasArgument("ShutterButton.allowshutterButton")) {
+      if (call.hasArgument("ShutterButton.allowShutterButton")) {
         builder.shutterButton.allowShutterButton = call.argument<Boolean>("ShutterButton.allowshutterButton")!!
       }
 
@@ -212,14 +211,19 @@ class KlippaScannerSdkPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
 
       if (call.hasArgument("DefaultColor")) {
         val default = call.argument<String>("DefaultColor")!!
-        if (default == "original") {
-          builder.colors.imageColorMode = KlippaImageColor.ORIGINAL
-        } else if (default == "enhanced") {
-          builder.colors.imageColorMode = KlippaImageColor.ENHANCED
-        } else if (default == "grayscale") {
-          builder.colors.imageColorMode = KlippaImageColor.GRAYSCALE
-        } else {
-          builder.colors.imageColorMode = KlippaImageColor.ORIGINAL
+        when (default) {
+            "original" -> {
+              builder.colors.imageColorMode = KlippaImageColor.ORIGINAL
+            }
+            "enhanced" -> {
+              builder.colors.imageColorMode = KlippaImageColor.ENHANCED
+            }
+            "grayscale" -> {
+              builder.colors.imageColorMode = KlippaImageColor.GRAYSCALE
+            }
+            else -> {
+              builder.colors.imageColorMode = KlippaImageColor.ORIGINAL
+            }
         }
       }
 
@@ -287,8 +291,72 @@ class KlippaScannerSdkPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
         builder.menu.userCanChangeColorSetting = call.argument<Boolean>("UserCanChangeColorSetting")!!
       }
 
-      val intent = builder.getIntent(context)
-      activityPluginBinding?.activity?.startActivity(intent)
+      val modes: MutableList<KlippaDocumentMode> = mutableListOf()
+
+      call.argument<HashMap<String, String>>("CameraModeSingle")?.let { cameraModeSingle ->
+        val singleDocumentMode = KlippaSingleDocumentMode()
+        cameraModeSingle["name"]?.let { name ->
+          singleDocumentMode.name = name
+        }
+        cameraModeSingle["message"]?.let { message ->
+          singleDocumentMode.instructions = Instructions(message, dismissHandler = {
+            singleDocumentModeInstructionsDismissed = true
+          })
+        }
+
+        modes.add(singleDocumentMode)
+      }
+
+      call.argument<HashMap<String, String>>("CameraModeMulti")?.let { cameraModeMulti ->
+        val multiDocumentMode = KlippaMultipleDocumentMode()
+        cameraModeMulti["name"]?.let { name ->
+          multiDocumentMode.name = name
+        }
+        cameraModeMulti["message"]?.let { message ->
+          multiDocumentMode.instructions = Instructions(message, dismissHandler = {
+            multiDocumentModeInstructionsDismissed = true
+          })
+        }
+
+        modes.add(multiDocumentMode)
+      }
+
+      call.argument<HashMap<String, String>>("CameraModeSegmented")?.let { cameraModeSegmented ->
+        val segmentedDocumentMode = KlippaSegmentedDocumentMode()
+        cameraModeSegmented["name"]?.let { name ->
+          segmentedDocumentMode.name = name
+        }
+        cameraModeSegmented["message"]?.let { message ->
+          segmentedDocumentMode.instructions = Instructions(message, dismissHandler = {
+            segmentedDocumentModeInstructionsDismissed = true
+          })
+        }
+
+        modes.add(segmentedDocumentMode)
+      }
+
+      if (modes.isNotEmpty()) {
+        var index = 0
+        call.argument<Int>("StartingIndex")?.let {
+          index = it
+        }
+        val cameraModes = KlippaCameraModes(
+          modes = modes,
+          startingIndex = index
+        )
+        builder.cameraModes = cameraModes
+      }
+
+      when (val intent = builder.build(context)) {
+        is Failure -> {
+          result.error(E_FAILED_TO_SHOW_SESSION, "Could not launch scanner session", intent.reason.message + "\n" + Log.getStackTraceString(intent.reason.cause))
+          return
+        }
+        is Success -> {
+          activityPluginBinding?.activity?.startActivity(intent.value)
+        }
+      }
+
 
       resultHandler = result
 
@@ -297,7 +365,7 @@ class KlippaScannerSdkPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
     }
   }
 
-  override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
+  override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
     channel.setMethodCallHandler(null)
   }